Organize leader stages into module
This updates components and experiment stage logic to ensure that
VoteForLeader and RevealVoted stages are added together (e.g., the
latter is linked to the former).

This also formats the "Add experiment stage" menu (in experiment
config builder) to include both standalone stages and modules;
and begins adding utils functions for the ranking game module
(includes surveys and chat connected to randomly-generated ranking
items).