Add ranking (Lost at Sea) game module to experiment config options
This enables experimenters to add the surveys/chat for Lost at Sea
(with the item pairs randomly generated and added to relevant stages).

Note that as of this commit, participants can advance through Lost at
Sea (ranking survey, chat) stages, but cannot yet view/update the data.
This functionality will be added in a follow-up commit.