Refactor chat components into stage component containing interface
Previously, the chat-interface component was used to render the entire
chat stage. This separates the interface (chat messages, input box) from
the overall stage to make it easier to implement different types of chat stages
(the current "basic" stage is just the interface and a footer, but we'll
want to create a Lost at Sea ranking chat stage that includes discussion
topics and the ability to advance through them).