Fix bug with participant access to all profiles
The Lit app was previously using `experimentService.participants`, which
is an extended set of profiles (with privateIds) meant for experimenters
and inaccessible to participants.

This commit switches over participant view references to profiles to use
the publicly-available profiles (under
`experimentService.experiment.participants`).

NOTE: This removes the temporary hack to fix advancing item discussions,
as it relied on access to private IDs. This means that advancing through
the items is currently broken—but will fix in a follow-up.