Convert RevealVoted stage to implicit Reveal stage
This renames the stage type/references from `RevealVoted` to `Reveal`
and updates logic so that a list of stages (to show results for) can be
added to the reveal stage.

This also redefines VoteForLeader as a basic (default) stage so that
the only game (collection of stages) is Lost at Sea.