Add reveal field (for showing results during reveal stage) to all stages
Right now, experimenters can only toggle reveal to true for LAS surveys
on the frontend; but this adds an optional reveal boolean to all stages
in case future stages require reveals.

This also adds the checkbox for showing/hiding reveal results for LAS
surveys and consequently (if the experimenter checks true)
displays a placeholder message for survey results in the reveal stage.
(Actual LAS survey results to be added in follow-up commit.)