Define experiment stages by unique IDs, not names
This adds an `id` field to experiment stages, as well as a `stageIds`
field to experiment that specifies the ordering of the stages.

This change eliminates reliance on stage numbering to make stage names
unique, which enables future editing of stage names as well as the
ability to specify stages during experiment configuration (e.g., for
composite stages).