Include pre-lobby stages in post-lobby experiments
This fixes the bug with not being able to add payout stages (as payout
stages may rely on pre-lobby survey stages).

Now that pre-lobby stages are included: for transfer participants,
set the current stage (in the post-lobby experiment) to the stage after
the current one (which should be the lobby stage).

NOTE: To properly transfer participants, we will need to copy over
private/public stage answers from the lobby experiment. (Left a TODO
in experiments.endpoints.)