Replace election multiple voting with drag-and-drop
NOTE: This does not yet implement logic to select current election
leader on the backend (instead, there's a TODO and temporary logic to
set the most recent voter as the leader).

NOTE: This allows participants to drag/drop profiles into the ranking zone,
but not reorder via drag/drop (instead, they must click the "close"
icon to remove the profile(s) from the ranking zone, then re-drag).
We'll add ability to reorder via drag/drop in a follow-up.