Fix template creation functionality
Update ExperimentTemplate to include fields necessary for template state
(e.g., whether or not to create multiple experiments).

We may eventually want to structure experiments under a Group
structure (with both lobby and experiment documents underneath) and
have templates mirror the Group structure. (This could also be useful
for adding a new experiment to a group post-creation?)