Add transfer stage view for participants
If timeout is enabled, the stage will render an overlay that prevents
participants from navigating away (e.g., to a previous stage) during
countdown.

If an experimenter is previewing the stage, or if the participant is
past the transfer stage, the countdown (if timeout enabled) will be
shown, but no action will happen when it reaches zero (i.e., the
participant will not be removed from the experiemnt).

Note that this stage does **not** include a stage footer because
the participant does not initiate progressing to the next stage; rather, they
progress when accepting the transfer (to be implemented).