Create index.html template for users to add Google Analytics
Instead of checking in `index.html`, this change checks in
`index.example.html` and requires users to save a fork as `index.html`
and fill in their Google Analytics ID (or comment out tracking).

This workflow is similar to copying over firebase configs; will clarify
in wiki documentation once merged to main.