Improve styling for sampling parameters.
Add a new name-value-input for custom body request fields. Move each numerical
sampling parameter into its own number-input div so they're not as wide.

Also set the minimum of the frequency penalty and presence penalty boxes to 0;
this makes them equal in width to the others, and it's unlikely that a user
would want to set them below 0. (If they really want to type a negative number,
I think it will work anyway.)