Remove redundant survey public data write from survey answer endpoint
This was added for automatic cohort transfer, but the trigger on
participant answer write should already do this. Let's see if the timing
can work without the double-write and if not, find a better workaround,
e.g., time delay in automatic transfer function to wait for trigger
function to catch up.

The reason we want the public data write in a trigger function is
because the participant survey answer may have been updated
programmatically (e.g., by an agent participant), not necessarily
through this endpoint.