Store participant assistance move on backend
This adds an endpoint for specifying assistance mode and uses the stored
mode to render the frontend UI. The new endpoint immediately queries for the
model response if relevant (i.e., advisor, delegate modes). Otherwise,
the user clicks a relevant button in the UI to call the existing
"request assistance" endpoint.

All relevant chip endpoints now write assistance response data to
participant answer under "current assistance" and/or "assistance
history."

Note that responses displayed in the UI still use temporary frontend
state variables and "coach" mode for responding to offers is not yet set
up in the frontend UI. These will be addressed in follow-up commits.