Write error chat messages for failed private chat triggers to continue conversation
This enables participants to continue responding if turn-taking is
enabled (as turn-taking logic checks whether the participant was the
last person to send a message; once an error chat message is written,
the participant will be able to send messages again).
If a participant is stuck waiting, they can click a "stop"
button to "cancel" the request.