Add frontend-only basic participant view to MultiAssetAllocationStage
This enables participants to adjust sliders for each of the stock
options and prevents them from moving on to the next stage if the
allocations do not sum to 100. No visualization, stock info, or
allocation "confirmation" yet. Answers are currently only cached on
the frontend, but a follow-up will add a backend endpoint to store
the participant answers.