Remove obsolete reality TV, salesperson templates
Reality TV template was initially intended to demonstrate agent
facilitation, but this is now possible via quickstart group chat
template (and we are moving towards a simplified model of agent
participants in the frontend UX).

Salesperson template was used to test a temporary salesperson stage,
but this stage is no longer supported.