Check for empty agent participant ID before attempting to fetch prompt
This was causing an issue (and breaking agent participants) when
experimenters explicitly specified an empty ID in order to use the
fallback prompt(s).

Also, add a constant for `DEFAULT_AGENT_PARTICIPANT_ID` in case we
decide to default to a specific agent configuration (rather than empty
ID --> fallback) when creating agent participants from dashboard.